Document_Analyzer/requirements.txt
numpy>=1.21.0
gensim>=4.1.2
transformers>=4.15.0
torch>=1.9.0
scikit-learn>=0.24.2
nltk>=3.6.3
spacy>=3.1.0
python-dotenv>=0.19.0
PyPDF2>=3.0.1
openai>=0.27.0
pandas>=1.3.0
